Add expired check for random item sampling
Summary:
Previously, we were not checking the expiration for the sampled item, which could give an incorrect
view of the item compositions. This could be more of a problem for items sampled from NVM cache
because NVM cache is deferring the evictions for expired items until the reclaims. This fixes the
issue by adding a check for expiration for item sampling for both DRAM and NVM items.

Reviewed By: udippant

Differential Revision: D52179872

fbshipit-source-id: b721f2b55fefd694ce84e99f6270f1de462a672e