Support for storage cache trace replay generator
Summary:
This change implements the replay generator (block-replay) for block storage cache. This change
assumes that the block storage cache caches the extent of fixed sized (e.g., 8MB) blocks of
distributed storage data in the unit of the chunks of given size (e.g., 256KB). E.g., if the chunk
size is 256KB, the full extent of the block of size 8MB will be cached as 32 cache items. The replay
generator also supports the sharded block storage, where the block of larger size is stored in each
storage node in sharded manner; i.e., if the full block of size 72MB could be sharded as 9 storage
node blocks (from shard_id 0~8), if the storage block size is 8MB.

Reviewed By: therealgymmy

Differential Revision: D52263052

fbshipit-source-id: 614f32e714c66c6e8e8ed7ab511a22ef3b951b70