dump usecase and sub-usecase id for memcache trace
Summary:
This change adds a feature to dump the usecase and sub-usecase id for memcache trace. The extraction
of the usecase is based on the following key format (https://fburl.com/wiki/aw9xa3wa); i.e., the
first field (with ":" as the separator) is used as the usecase.

<Usecase ID>:[<Subcategory>:]<ID Key>:[s<sevVersion>:]c<version>

There is one exception applied for key starting with "inference_platform", which seems to use
different key name schemes; the usecases of them are squashed to "inference_platform" for now.

In addition to the encoded usecase id, this change tries to emits the sub-usecase id. For
convenience, this change extracts the sub-usecase from the prefix list configerator config which is
used to set the TTL for memcache usecases (i.e., memcache/configs/default_config).

This change also adds the support for consistent key encoding with which key is encoded from the key
string instead of from linear mapping.

Reviewed By: therealgymmy

Differential Revision: D53028088

fbshipit-source-id: 61928f9977b39616f9873209ac376428e2920414