Transparent Item movements (#251)
Summary:
The new algorithm relies on the moving bit and does not require external synchronization. Data movement happens transparently for the client: if the client thread attempts to get a handle for the item being moved it will get a handle with wait context to wait till the movement is completed.

Pull Request resolved: https://github.com/facebook/CacheLib/pull/251

Reviewed By: jaesoo-fb

Differential Revision: D51402910

Pulled By: haowu14

fbshipit-source-id: 349933265fa8e888b5a99426bd7e15743b86f592