replaced the remaining use of folly::Baton instead of folly::fibers::Baton
Summary:
As discovered in T177120945, there is several places using folly::Baton instead of
folly::fibers::Baton mistakenly. The folly::Baton is unaware of fibers, so the wait() blocks whole
thread unintentionally. So, instead in the code that can be executed by the fibers, we should use
folly::fibers::Baton, which is little heavier, but supports the fiber waiters.

Reviewed By: alikhtarov

Differential Revision: D53629855

fbshipit-source-id: 401fbe9bd33593468bce45557d3a5d1e3dfadb91