run flush on reclaim thread
Summary:
As discovered in T177120945, it seems like there is some regression in p99 latency of insert
requests with navy-async. The reason for this might be the fact that we are running the flush in
line if issued for insert request on NavyThread. However, since the flush could take long, this
could cause unnecessarily high latency, particularly, if flushes are executed multiple times on the
same NavyThread as part of insert requests back-to-back. Since the flush can be executed async, this
change makes the flush to be run by the reclaim threads.

In doing so, this change also adds a feature to get the reference to the current NavyThread via
global function.

Reviewed By: haowu14

Differential Revision: D53491112

fbshipit-source-id: 309d43c9fc23343d006344661f618f03c9d30843