Fix a fiber/DMutex deadlock
Summary:
If getNextCandidate() is executed from within a fiber context, the `if (!markedForEviction)` code path can end up in the ~PutToken destructor. The token map is protected by a fiber-compatible TimedMutex which on contention preempts the current fiber. Since the DMutex is still owned by the preempted fiber, the next call on the same OS thread to acquire the DMutex will deadlock.

This is an instance of a general problem, where fiber-compatible mutexes must not be acquired within a critical section of non-fiber-compatible mutexes.

Reviewed By: therealgymmy

Differential Revision: D57882069

fbshipit-source-id: 611f6a97a125818e0253c4f8712c82b04e035a30