Fix bug in iteration logic that may return null handle
Summary:
Reported by user: https://fb.workplace.com/groups/cachelibusers/posts/2528004787386648

A couple of months ago we changed our eviction synchronization logic significantly. The outcome is that on lookup, we can return a miss if an item is concurrently marked as for eviction. We have missed updating the iteration logic to take this into account.

This diff fixes that and only returns valid handles.

Reviewed By: SLieve

Differential Revision: D58193077

fbshipit-source-id: 83331a700ca7aeb8006401375219892a139faf13