Adds documentation about FDP support in Cachelib (#308)
Summary:
This adds the relevant documentation related to the FDP integration to CacheLib which was merged in commit 009e89b.

This commit adds a separate page containing all the information about FDP and its use in CacheLib. Additionally, this page introduces a potential user of CacheLib with all the relevant steps to setup up an FDP SSD and to run CacheLib with FDP enabled.

The parameter to enable FDP in CacheBench "deviceEnableFDP" has been updated to the CacheBench configuration page.

The parameter to enable FDP in Navy Config
"navyConfig.setEnableFDP(enableFDP)" is updated to the Hybrid Cache configuration page.

Information on the FdpNvme class is added to the Navy Architecture Guide page.

Pull Request resolved: https://github.com/facebook/CacheLib/pull/308

Reviewed By: therealgymmy

Differential Revision: D58738320

Pulled By: haowu14

fbshipit-source-id: 0666c2df97c2817372c09beab8f48e20dc4821cd