Upload rebalancer configs in only a single column
Summary:
I messed up in D58759735 where each memory pool's rebalancer config is uploaded in a separate column. This led to our table becoming >2K columns! It's because across 100+ use-cases, people just end up with that many name combinations for pools.

The better way is to upload all rebalancer configs in a single column in a big json string.

Follow-up to this diff is to work with Scuba team to see how we can drop the old data so we go back to having fewer columns. Right now it's quite hard to navigate our table due to so many cols.

Differential Revision: D59233415

fbshipit-source-id: 790c6b051de58a3ef9f7376df0042712cdd845e9