Add ability to provision a memory pool on cache startup
Summary:
follow up to S424484

Add the ability for user to provision slabs for each AC in a memory pool. This is useful when a user knows their workload distribution in allocation sizes.

This would have prevented the issue in this sev which is due to large volume of large items request that quickly fills up the cache with large items on cold cache start up. Even with the rebalancer fix, it would take close to an hour for the rebalancer to move memory around. The best way for services like ucache would be to provision a memory pool with power-law distribution of slabs from small to large sizes.

Reviewed By: pbhandar2

Differential Revision: D58317640

fbshipit-source-id: 5eb676cabcacd52ea9b75d6d155ea0d19783bbef