Adding Unit Tests for FdpNvme class (#318)
Summary:
This commit adds the unit test cases for "FdpNvme" class.
It also adds a new constructor for FdpNvme. The existing constructor of FdpNvme communicates with the actual FDP device and makes the unit tests impossible to run without the actual device. So this commit adds a new constructor which allows to experiment with "FdpNvme" without the actual FDP device.

Pull Request resolved: https://github.com/facebook/CacheLib/pull/318

Reviewed By: therealgymmy

Differential Revision: D58702830

Pulled By: haowu14

fbshipit-source-id: b824c656611139d50bbc52796ceda1ed0bea693e