fix TTL bug in cachelib rust binding that sets all TTL > 1 to 1 and keeps TTL < 1
Summary:
The rust binding for LRU tries to convert all TTL less than 0 to 1.

Current bug is using ttl = min(user specified ttl, 1) which sets all user specified ttl > 1 to 1 and keeps ttl < 1 which is opposite of what we want.

We replaced it to set ttl to 0 when a 0 or None value is passed as duration. If a duration value > 0 and <= 1 second is provided, ttl is 1. If duration is greater than 1 second, then ttl is the closest second rounded down upto 2 * 30 * 86400 which is the total seconds in 2 months. So the current maximum allowed value of ttl is 2 months. When I tried 3 months, set failed. I believe this is because of the code here: https://www.internalfb.com/code/fbsource/[1b8f1f29260f]/fbcode/cachelib/allocator/CacheAllocator.h?lines=2671-2683.

Two 4 byte unsigned integers creation_time and ttlSecs are added to compute expiry time. A ttl value too large can cause it to overflow.

Reviewed By: therealgymmy

Differential Revision: D60137146

fbshipit-source-id: 35e8ef2a4d7be4e7251189e40129f103715dd4c0