rename addr in BlockCache::readEntry()
Summary:
To newcomer like me, it was a little confusing when I first read readEntry() code.
How about renaming 'addr' to 'addrEnd' as it's closer to the meaning and addrEnd is widely used in other places. (Even for the prototype of this function in [BlockCache.h](https://fburl.com/code/cd9k9myn))

Reviewed By: therealgymmy, pbhandar2

Differential Revision: D61724296

fbshipit-source-id: 4459dc51beb34f8f465e1f45f849c4c2a345082e