Fix a fiber/DMutex deadlock, take 2
Summary:
If getNextCandidate() is executed from within a fiber context, the `if (!markedForEviction)` code path can end up in the `~PutToken` destructor. The token map is protected by a fiber-compatible `TimedMutex` which on contention preempts the current fiber. Since the `DMutex` is still owned by the preempted fiber, the next call on the same OS thread to acquire the `DMutex` will deadlock.

This is an instance of a general problem, where fiber-compatible mutexes must not be acquired within a critical section of non-fiber-compatible mutexes.

We fix it by running the `if (!markedForEviction)` path inline with the Token acquisition, under the original `try_lock()`.

Reviewed By: therealgymmy, haowu14

Differential Revision: D58144623

fbshipit-source-id: e849a2e1b1509e5d58892cdcaa024848ad65f6e9