Add a 5 byte compressed pointer for t2
Summary:
This diff adds 5-byte compressed pointer for T2 support.

It also adds a benchmark to explore further optimizations to 5-byte pointer to improve performance or to encode additional information about an item in the future.

However, note that CacheAllocator using 5-byte compressed pointer can only map 256GB of data. This is because the free allocation list in allocation classes use 4-byte compressed pointer. Changing this to a template type leads to a large number of complex changes in the codebase. All changes up to here should be a no-op change for current customers.

The next diff will set the 5-byte compressed pointer to be the default for the free list in allocation classes. It does not bloat the item size for caches <256GB as the item still uses 4-byte compressed pointer by default. This allows us to avoid complicated changes to the codebase with a minimal change to current users. CacheLib will support the new CacheAllocators with 5-byte compressed pointer which supports caches up to 32TB.

This diff also saves 1 bit for identifying where the compressed pointer belongs to an item in DRAM or NVM. The implementation is due.

This diff splits declaration of allocators in CacheAllocator.cpp into separate files which helps with build time.

Reviewed By: alikhtarov

Differential Revision: D62625822

fbshipit-source-id: 337b7cbcbeb7a8b68d124277ecbe97d1d850afca