Retry reading after checksum error
Summary:
For S421120, we're thinking that transient checksum errors are happening.
With D58986608, re-read was added for debugging purpose, and we could see that most logs (all the cases that were investigated) show that re-read succeeded. So this checksum error may be due to the race between read/write for the same LBA.
Regardless of the root cause (we're suspecting lower layer including kernel/ device driver or SSD itself), we can mitigate the SEV by returning the result of re-read if it was successful.

A counter for this case (navy_bc_retry_reads) was also added for monitoring purpose

Reviewed By: haowu14

Differential Revision: D63830787

fbshipit-source-id: ea207526d74750d8591928f100c09a60251eb631