Add util to enable encryption by items
Summary:
Add the util for encryption by items.

Currently, BigCache uses CacheLib to encrypt data in 4k blocks when we write data to flash. However, in our project to preserve cache on reboot, we will rotate encryption keys and need to remember which one has been used for each item so that we can decrypt it successfully. Encryption by blocks does not work well.

This diff adds Add the util for encryption by items.

Reviewed By: haowu14

Differential Revision: D63960128

fbshipit-source-id: 40b773f2b0da810addbf146ed7d884f118d1c85e