Adding error status for ChecksumError
Summary:
Currently in cachelib, checksum error is treated as DeviceError.
This only makes sense when device can silently return invalid data without returning real error code. However, all the SSDs will have its own ECC check and multiple stages of protection to detect/correct corrupted data. So in most cases, data won't be corrupted by device and when it really happens, SSD will return UNC (Uncorrectable) error instead of silently returning corrupted data. If device is returning wrong but not corrupted data (S421120), this is not real device error but more like the problem with sw (on host if it's a bug from cachelib) or fw (from ssd, if it's a bug from the ssd).

Based on this, we need to consider most of checksum errors as different ones from real device error.
After the fix for S421120 (retry reading with the checksum error), we could see multiple cases that device is not really  responding (for any config reasons or when the host is not in good status, etc.) and this real error with the device is still treated as checksum error, since we don't really differentiate it, so we need to add the error status for the checksum error (which will be mostly error cases with device still working and returning data).

This diff adds new error code (ChecksumError) to differentiate checksum error from the real device error.
(Regarding the previous workaround for S421120, we will retry reading only for checksum error case after this diff)

Reviewed By: haowu14

Differential Revision: D67266835

fbshipit-source-id: 719e61928cca342b1e6458d317cf87c5ec4275b0