Adding OCP plugin support for getting NAND write bytes
Summary:
NOTE: **Update w/ V3**
Added resolvedMap for the model number it already found how to get NAND write bytes info.
In addition to addressing Hao's suggestion, it will also improve the efficiency, since it will not search for every registered vendor name with the model number to find the matching entry from the vendor map once a model name is resolved. (Currently, it will search for matching vendor name for every entries per each stat collection - default every 1 min)

With Cachebench, it will gather stats for every 1 minutes and # of NAND written bytes is one of the stats retrieved.
Since this info is only accessible by vendor specific command and format, current code will determine the vendor by model, then use the vendor specific function (which will use vendor specific plug-in for NVMe CLI) accordingly. (https://fburl.com/code/lm2arxeo)

With new OCP spec, this will change with the new device (or FW).
This is tricky, since new 'OCP way supporting device' may not report the stat with the old way (vendor specific command and format), and it's not determined by model name, since same model may or may not support OCP way (by FW, since not every device is for OCP adoption companies).
Also, with new OCP way, it's unified way and new device model (supporting this) doesn't need to be added to the map we are using for vendor specific function, since new devices will support the same OCP way eventually.
The other thing to consider is, we will have both OCP supporting and non-supporting devices for a while.

This diff will cover all the cases to use OCP way to retrieve NAND written bytes considering the situation for now.
1. For the devices we already have in vendor map, if it's failing to retrieve the info with the vendor specific way, it will try with the OCP way to see if it supports the new way
2. For the devices we don't have in vendor map, current code will fail and throw the error. Instead of this, it will try with the new OCP way to see if it supports the new way.
In both cases, if it succeeds with the OCP way, it will change the vendor map to use OCP way for this run cycle.
Though, it will do trial-and-error for the first check, this is not performance critical path anyway, so it'll be fine.

Reviewed By: haowu14

Differential Revision: D67221896

fbshipit-source-id: 5b2c4f4dd7189de8a14e09120b3f1f26c2366360