Don't disable navy with the checksum error (1/2)
Summary:
When we differentiate the checksum error from the real device error, there's no reason to disable navy when we get checksum error. We can just remove the corrupted entry from the cache.
For real device error, it will be still good idea to disable it, since device is not working properly and not returning data. (Maybe we can add retries before we disable the navy in the future)

This diff doesn't change the logic to disable navy with the checksum error just yet.
- It will just separate the test cases and mock device's bad status (IoReqFailure, DataCorruption).
- ChecksumError from BlockCache will be just returned to caller instead of being changed to BadState error.

Reviewed By: haowu14

Differential Revision: D67308028

fbshipit-source-id: 496533b6ced337658664d09685fc350e16b4b625