Don't disable navy with the checksum error (2/2)
Summary:
When we differentiate the checksum error from the real device error, there's no reason to disable navy when we get checksum error. We can just remove the corrupted entry from the cache.
For real device error, it will be still good idea to disable it, since device is not working properly and not returning data. (Maybe we can add retries before we disable the navy in the future)

This diff will do change the logic to avoid disabling navy when it got checksum error.

Reviewed By: haowu14

Differential Revision: D67434921

fbshipit-source-id: 911c870d408db5b2a4258c9e87dd412f06c41bc0