rename disableNvmCacheOnBadState_S421120 to disableNvmCacheOnBadState
Summary:
We are now differentiating checksum error from the real device error and not disabling navy for checksum error, this disableNvmCacheOnBadState_S421120 (which was added to work around S421120 originally) is not used.

Since this could be useful in the future for debugging/testing and in case when we should not disable navy in any error cases, it'll be good idea to keep it, so I'm just changing the name by removing _S421120 tag.

Reviewed By: haowu14

Differential Revision: D67346246

fbshipit-source-id: 950c7a15708284c0f31d354641c47d8f74d2eeba