Fix Variable Names for P100(Max) Latency in Cachebench
This commit fixes a bug where the P100 Latency is not displayed in the output of Cachebench. The cause of the bug was due to different variable names used for P100 Latency in Cachelib and Cachebench, and the variable name in Cachebench has been modified to match the variable name used in Cachelib.