Adjust the ChainedHashTable assesment to account for 4KB alignment.
Summary:
We will be changing the size of compressed pointer from 4 byte to 5 byte.

The problem that arises is that the size of the hash table is no longer 4KB aligned which means the size of the shared memory when we persist the hash table will be larger than the size of hash table to ensure 4KB alignment.

When we load the data in shared memory as a hash table, the size of  hash table based on the number of entries will not match the size of the shared memory.

The size of hash table is computed as

    size_t size() const noexcept {
      return numBuckets_ * sizeof(CompressedPtrType);
    }

The value of numBuckets_ has to be a power of 2 and the value for power is defined by the user. Usually we use 10, so the numBuckets_ is 1024.

This means with a 4 byte compressed pointer, the total size of shared memory and hash table would be 4096 bytes. Even if we keep increasing the power value to 11, 12 and so on, the resulting size is always 4KB aligned. When we persist, the size of shared memory required to store the hash table would also be 4096 bytes.

When we increased the size to compressed pointer to 5 byte, the size of hash table based on number of buckets is 5120 bytes. The shared memory that we need now is 8192 bytes to ensure 4KB alignment.

Reviewed By: haowu14

Differential Revision: D68539351

fbshipit-source-id: 903c4df84fbf730526e43d0cb843a7775e50f2ed