Don't try to copy TimedMutex
Summary: Copy constructor of `TimedMutex` is deleted. The overload used to construct this vector here is using the copy constructor. I don't know why `libstdc++` allowed it and it's likely not correct. Since we are just default constructing a `TimedMutex`, we can omit the second argument and make it just default initialize the mutexes.

Reviewed By: florazzz

Differential Revision: D71567619

fbshipit-source-id: 21a728e6137c4af22269d5bb0533db2abe8de875