Add a config to issue flush async from region manager's worker threads
Summary:
Before this diff, when a reclaim triggers a flush, that flush will be executed synchronously in the reclaim thread. In this diff, a config is introduced to schedule the flush.

Doing the flush inline during a reclaim can make the region allocator lock being held for the entire duration of the flush (doing pwrite for sync io, issuing a device write for asyncio). The same region allocator lock is required by nvm insertion. So this could cause high nvm.insert.latency. We observe this in bigcache.

If we schedule the flush, we could run out of in-memory buffer since now the flush for this current region will only be flushed after this run of reclaim finishes. But we can avoid this by assigning a larger number of clean regions.

This change has been validated on both bigcache and ucache and observed lower reclaim time spent and reinsertion errors.

Put up this diff to land it first. Once unit test is added, this config should be removed and a flush should always run async from the worker threads.

Reviewed By: byahn0996

Differential Revision: D72364726

fbshipit-source-id: 2fa8e18fd22bd65fe6e7b791a14887c3c3330238