Simple implementation of FixedSizeIndex
Summary:
The first version of FixedSizeIndex implementation. It's mostly based on Anton's 5B prototype/hack code.
There will be many improvement points here and there, but will be good to have a pure minimum version as a starting point to make it work and evaluate with the different options easily.

Reviewed By: alikhtarov

Differential Revision: D73440703

fbshipit-source-id: 6ef189d8df8bc59bf3aa21fa13d2abcdc2fb4ebb