Replace unnecessary index->lookup with index->peek call
Summary:
In BC, the difference with index's lookup() from peek() is whether we increase hit count or not.
We currently have index->lookup() calls which are not needed and falsely increasing current/total hit counts in index.

This diff simply changes those index->lookup() call to index->peek().
(We have one more place in BC::remove() => BC::lookup() => index->lookup() path, This is not touched for now, since it needs more changes and also not important, since item is about to be removed anyway).

Reviewed By: haowu14

Differential Revision: D75322823

fbshipit-source-id: 4d4d61c684a5cd83aefbdf7939e5c4831d9dccd3