Add ability to clone Git repo branches
Summary:
Unfortunately we can't just use `git.rev` in manifests to specify branches as we clone using `--depth`. This means `rev-parse` on a remote branch name will fail.

This diff adds `git.branch` which allows us to be explicit when we want a branch and so check this out in the first place. By doing this future uses of `rev-parse` will also understand the branch name.

Reviewed By: martindemello

Differential Revision: D79660472

fbshipit-source-id: 6b04158bdd72c28864322be28d88ee4e41f54bc8