Allow arbitrary size keys
Summary:
Lift the 255 byte limitation on keys.  Since the majority of keys are still going to be optimized for <= 255 bytes, use a small key optimization:

- If key is <= 255 bytes, keep the same packed format in use today (8 bits for key size, 24 bits for value size)
- If key is > 255 bytes, set the packed key size field to 0 and embed an extra 4 bytes before the beginning of the key with the actual key size

Note: key sizes are limited to the same size as value sizes (16MB) even though we could support storing up to 4GB.  It's strange for a key to be bigger than the data and there's no reason for gigantic keys.

Differential Revision: D78276540

fbshipit-source-id: a8a232756353ff479dd7945bb7e2066d1d17f50b