Reduce fill lock contention in NvmCache hot miss path
Summary:
Make FillLock a RW lock. Only shared lock is needed to check if the item exists in RAM, in flight, or to call Navy's couldExist(). Then fallback to the exclusive lock as before.

Use TimedRWMutexWritePriority to avoid readers starving writers: if we have a request waiting to fill from NVM, always allow it to succeed first before queueing up additional readers.

Reviewed By: rlyerly, AlnisM

Differential Revision: D81735547

fbshipit-source-id: 3185f7bc3de5af4812232c9e83018ef77cccfae2