Reduce lock contention in InFlightPuts
Summary:
invalidateToken() is called on the hot miss path in NvmCache::find(). We can avoid the exclusive lock by changing the valid state to an atomic, as the only transition is true -> false.

Use TimedRWMutexWritePriority to avoid readers starving writers: if we have a request waiting to fill from NVM, always allow it to succeed first before queueing up additional readers.

Reviewed By: AlnisM

Differential Revision: D81963166

fbshipit-source-id: e1c3d33bfc5b0a5f008e68d1b694997329921608