FixedSizeIndex: Handling non-existing key case
Summary:
With FixedSizeIndex, there's more possibilities for false positive
For the non-existing item, it may try best effort to find it, and may return false positive PackedItemRecord instance.
But when we are sure that it doesn't really exist, it's better to return "it's not there".

To simply indicate it doesn't exist, I changed recordRef() to recordPtr() in ExclusiveLockedBucket and SharedLockedBucket. Those members were already pointer type within the structures, so there's no increased vulnerability by using pointer type directly.

Reviewed By: rlyerly

Differential Revision: D77402423

fbshipit-source-id: ed65de8f94f724cbc86835320c0aac6b6de47d26