Create ShmManager for navy
Summary:
To use shm for persistence in BlockCache, creating ShmManager for navy using the path under {main_cache_dir}/navy_shm

Update (with Version 5)
Using NavyConfig to pass down the necessary info to create ShmManager doesn't seem to be good idea. By using NavyConfig, users can also call it to tweak the values. We don't want to give the option to alter it. It just needs to be the same value we use for main cache (which is already configured by user).
Instead of having it in NavyConfig, it's added as a parameter (NavyShmParams) to create navy and will be given with the same value used for main cache.

Reviewed By: rlyerly

Differential Revision: D81619566

fbshipit-source-id: 8607136c3b2d01bd2518fc642d979210005b8eb0