Generic CacheItem & Handles
Summary:
Create generic CacheItems that are maintained by individual cache components (e.g., RAM cache).  Concrete implementations will need to implement this interface; this is how users will interact with the actual items.  This will probably get extended as we start concretizing it for RAM + flash cache components.

Also create Handles that wrap these CacheItems and that can be passed back to users.  The handles should be generic enough to be used across all concrete implementations; we shouldn't be re-implementing the same refcounting logic for every cache component.

We need a tiny amount of CacheComponent to correctly free items when refcounts drop to zero.

Reviewed By: AlnisM

Differential Revision: D82849311

fbshipit-source-id: 7c8710fcc36858d84ed0c3d4bfc72bf0e00ec489