Add more metrics to Objectcache
Summary: Add more metrics to object cache. It has been an issues during SEVs and debugging. We added monitoring for all metrics that are relevant for cache resizing such as free mem, rss bytes, number of entries in the cache and the current limit on the number of entries in the cache.

Reviewed By: AlnisM

Differential Revision: D84526808

fbshipit-source-id: d0d4158500c87610b5636931d6696cf39170651a