Generic CacheComponent
Summary:
Add the CacheComponent interface to be implemented by different components like RAM cache or Navy flash engines.  Again, we'll likely need to extend over time but this is enough to get us started.

Note that one of the principles of this interface is to *not* throw exceptions.  Instead we'll returned structured error types.  This will:

1. Make all error handling explicit
2. Avoid the perf overhead of throwing

We also are trying to tighten up some of the API semantics and use the type system to enforce some invariants.  In particular, we're differentiating between allocated-but-not-inserted items (AllocatedHandle) and allocated-AND-inserted items (Read/WriteHandle).  This makes it more clear for the user what the current status is at any point in the code.

Ideally we'd like to enforce the invariant that if you have a handle then you've got a reference to a valid item.  We can't exactly get there due to needing to support move semantics, so instead we make moved-out handles unusable (similarly to the existing implementation).  However we're making some of the APIs take rvalue-refs to handles to indicate that after the API call (successfully) completes, the handle is no longer usable.

Reviewed By: AlnisM

Differential Revision: D83660014

fbshipit-source-id: 1a569ff8427d694d829e575a76aed0828f1aa781