abort slab release if taking too long
Summary:
Slab rebalancing can sometimes take an excessively long time (e.g., longer than 1 hour) when the slab selected for release contains an item with a reference count > 0. In these cases, the rebalance operation is blocked until the item is released. By default, slab rebalancing is triggered every second, but a new rebalance cannot begin until the previous one completes. This means that if a rebalance gets stuck, all subsequent rebalances are blocked.

To address this, I have added an option to provide a timeout for slab release. If a slab release takes longer than this timeout, the operation is aborted.

The slab to evict is chosen randomly, so aborting the rebalance operation allows to attempt eviction of a different slab.

Reviewed By: rlyerly

Differential Revision: D82880706

fbshipit-source-id: 9e5affe9c0cdc3f70f8b8c7d098be17d835ebceb