add alloc size helper function for minimizing internal/external fragmentation
Summary:
Adding a helper function which can be used to optimize allocation classes and reduce fragmentation.

Example: We know that all of our items are between 42384 bytes and 103496 bytes. Slab size is 2^22.

2^22 / 42384 = 98.95
2^22 / 103496 = 40.52
This means that if we used 42384 and 103496 as allocation classes, we would be able to store 98 and 40 allocations per slab. Between these 2 allocation sizes, there 58 allocation sizes that result in a unique number of allocations per slab, i.e., we can use 2^22 / 40, 2^22 / 41, ..., 2^22/98 as our allocation sizes.

The function that I am adding is making sure that the allocation sizes are aligned and minimize external fragmentation.

Reviewed By: rlyerly

Differential Revision: D85119686

fbshipit-source-id: 0081bbfb6761816a628d49dadc53ca9f0205f36a