Fixing underflow bug with unit test
Summary:
reclaimsOutstanding_ will be underflowed to 0xFFFFFFFF when startReclaim() is manually called for unit test. (This variable will be updated to >0 value from getCleanRegions() before it's decreased after reclaim(), but by manually calling startReclaim(), it's 0 before decreased)
With this bug with the unit tests code path, getCleanRegions() can't trigger next reclaim. Rather than fixing this problem, current code relies on this incorrect behavior, so this diff will fix this.

Reviewed By: rlyerly

Differential Revision: D84429300

fbshipit-source-id: 295db8dfda1a64203407113e9460f6422af8eadc