Add a new mode that only uses free memory to adjust cache size
Summary:
Adding a new memory mode, ObjCacheSizeControlMode::FreeMemoryOnly, for testing. It makes sure every time the object size controller runs, the free memory in the system is above the lower limit. It gently expands the cache on each iteration if the cache has more free memory than the lower limit.

We can always expand the cache more aggressively but for now focusing on making sure the system memory usage stays above the lower limit.

We also shrink the cache during insertion. Size controller is run periodically but async adjustment might not be enough to prevent OOMs. So adding a check during insertion as well.

Reviewed By: AlnisM

Differential Revision: D85194675

fbshipit-source-id: 2f84f59c1e96e4bbb3359d5d5f799535bc63e9ad