update hits in HitsPerSlab even when no slab rebalancing occurs
Summary:
Currently, if HitsPerSlab decides not to perform slab rebalancing, it will not update hits for each allocation class. In one case, I was seeing that no slab rebalancing was happening for hours and this makes it hard for an allocation class to catch up. Even if it receives a lot of hits in one second, it will not get a slab.

I think this is especially bad when maxLruTailAge or classIdTargetEvictionAge are used. These configs allow one to prioritize specific allocation classes as a victim, even though they can have higher hits per slab than the chosen receiver class. This might disable slab rebalancing for a while and make hits accumulate without updateHits ever getting called.

In this diff, I add an option to change this behavior such that updateHits is always called regardless of whether a slab is actually rebalanced or not.

Might want to enable this by default but I am not sure about all the effects this wil have yet.

Reviewed By: pbhandar2

Differential Revision: D86054694

fbshipit-source-id: 0a24f6f5f95a0b5e17b977aeefe8723ca7f43846