provide option to pick victim based on alloc size
Summary:
Currently, memory monitor uses the configured rebalancer strategy to decide which class to evict a slab from. If the strategy does not pick a victim, memory monitor falls back to picking the class with the most slabs.

Memory monitor is trying to reduce memory pressure but since we still need to keep the index of the items inserted into block cache in DRAM, it is better to choose larger allocation sizes which contain fewer items per slab.

Picking a larger allocation allocation size also makes it less likely for cachelib to get stuck on slab release waiting for an item to have not references.

In this diff I introduce a new heuristic for picking a victim allocation class. The idea is that we pick randomly one of the x largest allocation sizes which have at least y slabs. The default strategy is still the total slabs based one but users can configure which one to use.

Reviewed By: pbhandar2

Differential Revision: D85471620

fbshipit-source-id: 40cc3becc5bcc74757196db80ef0e39466e7ec3b